fix: harmonize search signatures and unify filter value types
- Updated MainWindow and QuickWindow to use QVariant for filter values.
- Ensured DatabaseManager definitions match header signatures exactly.
- Added missing QVariant include in DatabaseManager.h.
- Fixed FilterPanel to use QVariant directly in stats updates.

+++ src/ui/FilterPanel.cpp
@@ -223,7 +223,7 @@ void FilterPanel::updateStats(const QString& keyword, const QString& type, const
         }
     }
 
-    QVariantMap stats = DatabaseManager::instance().getFilterStats(keyword, type, value.toInt());
+    QVariantMap stats = DatabaseManager::instance().getFilterStats(keyword, type, value);
 
     auto addStatsToRoot = [&](const QString& key, const QString& iconName, const QString& iconColor) {
         if (!m_roots.contains(key)) return;
